refactor(parser): extract raw YAML schema types into raw module
Refactor the parser by moving raw serde-deserializable types and related
functions that reflect the YAML schema into a new `raw` module. This cleanly
separates the raw representation from the typed model and parsing logic.

Additional changes:
- Update parser to use types and functions from the new raw module.
- Add validation helper functions for join and taint headers.
- Remove deprecated code and TryFrom implementations from the parser.
- Add parameterized tests for schema validation errors using rstest.

This improves code organization and maintainability by clearly separating
schema mirroring types from parsing and conversion logic.
