+++ crates/sempai-yaml/src/parser.rs
@@ -1,14 +1,15 @@
-//! Parser entrypoints and raw schema conversion for `sempai_yaml`.
+//! Parser entrypoints and rule construction logic.
 
-use serde::Deserialize;
-use serde_json::Value;
 use serde_saphyr::Spanned;
 
 use sempai_core::{DiagnosticCode, DiagnosticReport, SourceSpan};
 
 use crate::model::{
-    ExtractQueryPrincipal, LegacyClause, LegacyFormula, LegacyValue, MatchFormula, Rule, RuleFile,
-    RuleMode, RulePrincipal, RuleSeverity, SearchQueryPrincipal, TaintQueryPrincipal,
+    ExtractQueryPrincipal, LegacyFormula, Rule, RuleFile, RuleMode, RulePrincipal,
+    SearchQueryPrincipal, TaintQueryPrincipal,
+};
+use crate::raw::{
+    parse_mode, parse_severity, schema_error, singleton_formula, RawRule, RawRuleFile,
 };
 use crate::source_map::SourceMap;
 
@@ -31,105 +32,6 @@ pub fn parse_rule_file(yaml: &str, source_uri: Option<&str>) -> Result<RuleFile,
     Ok(RuleFile::new(rules))
 }
 
-#[derive(Debug, Deserialize)]
-#[serde(deny_unknown_fields)]
-struct RawRuleFile {
-    rules: Vec<RawRule>,
-}
-
-// Allow unknown fields for forward compatibility with future Semgrep rule extensions.
-// The official Semgrep schema sets `additionalProperties: true` on the rule object.
-#[derive(Debug, Deserialize)]
-struct RawRule {
-    id: Option<Spanned<String>>,
-    message: Option<Spanned<String>>,
-    languages: Option<Spanned<Vec<String>>>,
-    severity: Option<Spanned<String>>,
-    mode: Option<Spanned<String>>,
-    #[serde(rename = "min-version")]
-    min_version: Option<Spanned<String>>,
-    #[serde(rename = "max-version")]
-    max_version: Option<Spanned<String>>,
-    pattern: Option<Spanned<String>>,
-    #[serde(rename = "pattern-regex")]
-    pattern_regex: Option<Spanned<String>>,
-    patterns: Option<Spanned<Vec<RawLegacyClause>>>,
-    #[serde(rename = "pattern-either")]
-    pattern_either: Option<Spanned<Vec<RawLegacyFormulaObject>>>,
-    #[serde(rename = "match")]
-    match_formula: Option<Spanned<RawMatchFormula>>,
-    #[serde(rename = "dest-language")]
-    dest_language: Option<Spanned<String>>,
-    extract: Option<Spanned<String>>,
-    join: Option<Spanned<Value>>,
-    taint: Option<Spanned<Value>>,
-    #[serde(rename = "pattern-sources")]
-    pattern_sources: Option<Spanned<Value>>,
-    #[serde(rename = "pattern-sanitizers")]
-    pattern_sanitizers: Option<Spanned<Value>>,
-    #[serde(rename = "pattern-sinks")]
-    pattern_sinks: Option<Spanned<Value>>,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(untagged)]
-enum RawLegacyClause {
-    Formula(RawLegacyFormulaObject),
-    Constraint(Value),
-}
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(untagged)]
-enum RawLegacyValue {
-    String(String),
-    Formula(Box<RawLegacyFormulaObject>),
-}
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(deny_unknown_fields)]
-struct RawLegacyFormulaObject {
-    pattern: Option<String>,
-    #[serde(rename = "pattern-regex")]
-    pattern_regex: Option<String>,
-    patterns: Option<Vec<RawLegacyClause>>,
-    #[serde(rename = "pattern-either")]
-    pattern_either: Option<Vec<RawLegacyFormulaObject>>,
-    #[serde(rename = "pattern-not")]
-    pattern_not: Option<RawLegacyValue>,
-    #[serde(rename = "pattern-inside")]
-    pattern_inside: Option<RawLegacyValue>,
-    #[serde(rename = "pattern-not-inside")]
-    pattern_not_inside: Option<RawLegacyValue>,
-    #[serde(rename = "pattern-not-regex")]
-    pattern_not_regex: Option<String>,
-    #[serde(rename = "semgrep-internal-pattern-anywhere")]
-    anywhere: Option<RawLegacyValue>,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(untagged)]
-enum RawMatchFormula {
-    String(String),
-    Object(Box<RawMatchFormulaObject>),
-}
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(deny_unknown_fields)]
-struct RawMatchFormulaObject {
-    pattern: Option<String>,
-    regex: Option<String>,
-    all: Option<Vec<RawMatchFormula>>,
-    any: Option<Vec<RawMatchFormula>>,
-    not: Option<Box<RawMatchFormula>>,
-    inside: Option<Box<RawMatchFormula>>,
-    anywhere: Option<Box<RawMatchFormula>>,
-    #[serde(rename = "where")]
-    where_clauses: Option<Vec<Value>>,
-    #[serde(rename = "as")]
-    as_name: Option<String>,
-    fix: Option<String>,
-}
-
 fn build_rule(
     raw: RawRule,
     index: usize,
@@ -146,7 +48,7 @@ fn build_rule(
         rule_span.clone(),
         "add a stable rule id",
     )?;
-    let mode = RuleMode::from_optional(raw.mode.as_ref().map(|mode| mode.value.as_str()));
+    let mode = parse_mode(raw.mode.as_ref().map(|mode| mode.value.as_str()));
     let min_version = raw.min_version.clone().map(|value| value.value);
     let max_version = raw.max_version.clone().map(|value| value.value);
 
@@ -175,19 +77,6 @@ fn build_rule(
         max_version,
         principal,
     })
-}
-
-fn singleton_formula<F>(
-    mut formulas: Vec<LegacyFormula>,
-    make_error: F,
-) -> Result<LegacyFormula, DiagnosticReport>
-where
-    F: FnOnce(usize) -> DiagnosticReport,
-{
-    match formulas.len() {
-        1 => Ok(formulas.remove(0)),
-        len => Err(make_error(len)),
-    }
 }
 
 fn validate_search_header(raw: &RawRule, span: Option<SourceSpan>) -> Result<(), DiagnosticReport> {
@@ -237,6 +126,50 @@ fn validate_extract_header(
     Ok(())
 }
 
+fn validate_join_header(
+    raw: &RawRule,
+    span: Option<SourceSpan>,
+) -> Result<(), DiagnosticReport> {
+    require(
+        raw.message.clone(),
+        "message",
+        span.clone(),
+        "add a rule message explaining the match",
+    )?;
+    require(
+        raw.severity.clone(),
+        "severity",
+        span,
+        "choose a schema-aligned severity such as WARNING or ERROR",
+    )?;
+    Ok(())
+}
+
+fn validate_taint_header(
+    raw: &RawRule,
+    span: Option<SourceSpan>,
+) -> Result<(), DiagnosticReport> {
+    require(
+        raw.message.clone(),
+        "message",
+        span.clone(),
+        "add a rule message explaining the match",
+    )?;
+    require(
+        raw.languages.clone(),
+        "languages",
+        span.clone(),
+        "declare at least one target language",
+    )?;
+    require(
+        raw.severity.clone(),
+        "severity",
+        span,
+        "choose a schema-aligned severity such as WARNING or ERROR",
+    )?;
+    Ok(())
+}
+
 fn build_search_rule(
     raw: &RawRule,
     rule_span: Option<SourceSpan>,
@@ -279,20 +212,8 @@ fn build_join_rule(
     raw: &RawRule,
     rule_span: Option<SourceSpan>,
 ) -> Result<RulePrincipal, DiagnosticReport> {
-    let message = require(
-        raw.message.clone(),
-        "message",
-        rule_span.clone(),
-        "add a rule message explaining the match",
-    )?;
-    let severity = require(
-        raw.severity.clone(),
-        "severity",
-        rule_span.clone(),
-        "choose a schema-aligned severity such as WARNING or ERROR",
-    )?;
+    validate_join_header(raw, rule_span.clone())?;
     let join = require(raw.join.clone(), "join", rule_span, "add a join definition")?;
-    drop((message, severity));
     Ok(RulePrincipal::Join(join))
 }
 
@@ -300,25 +221,7 @@ fn build_taint_rule(
     raw: &RawRule,
     rule_span: Option<SourceSpan>,
 ) -> Result<RulePrincipal, DiagnosticReport> {
-    let message = require(
-        raw.message.clone(),
-        "message",
-        rule_span.clone(),
-        "add a rule message explaining the match",
-    )?;
-    let languages = require(
-        raw.languages.clone(),
-        "languages",
-        rule_span.clone(),
-        "declare at least one target language",
-    )?;
-    let severity = require(
-        raw.severity.clone(),
-        "severity",
-        rule_span.clone(),
-        "choose a schema-aligned severity such as WARNING or ERROR",
-    )?;
-    drop((message, languages, severity));
+    validate_taint_header(raw, rule_span.clone())?;
 
     if let Some(taint) = raw.taint.clone() {
         return Ok(RulePrincipal::Taint(TaintQueryPrincipal::New(taint.value)));
@@ -425,19 +328,6 @@ fn build_legacy_principal(
     })
 }
 
-fn parse_severity(
-    value: &Spanned<String>,
-    fallback_span: Option<&SourceSpan>,
-) -> Result<RuleSeverity, DiagnosticReport> {
-    RuleSeverity::parse(&value.value).ok_or_else(|| {
-        schema_error(
-            format!("unsupported severity `{}`", value.value),
-            fallback_span.cloned(),
-            "use one of ERROR, WARNING, INFO, INVENTORY, EXPERIMENT, CRITICAL, HIGH, MEDIUM, or LOW",
-        )
-    })
-}
-
 fn require<T: Clone>(
     value: Option<Spanned<T>>,
     field: &str,
@@ -476,218 +366,4 @@ const fn is_schema_error(error: &serde_saphyr::Error) -> bool {
             | serde_saphyr::Error::SerdeMissingField { .. }
             | serde_saphyr::Error::DuplicateMappingKey { .. }
     )
-}
-
-fn schema_error(message: String, span: Option<SourceSpan>, note: &str) -> DiagnosticReport {
-    DiagnosticReport::validation_error(
-        DiagnosticCode::ESempaiSchemaInvalid,
-        message,
-        span,
-        vec![note.to_owned()],
-    )
-}
-
-impl TryFrom<RawLegacyClause> for LegacyClause {
-    type Error = DiagnosticReport;
-
-    fn try_from(value: RawLegacyClause) -> Result<Self, Self::Error> {
-        match value {
-            RawLegacyClause::Formula(formula) => Ok(Self::Formula(formula.try_into()?)),
-            RawLegacyClause::Constraint(constraint) => Ok(Self::Constraint(constraint)),
-        }
-    }
-}
-
-impl TryFrom<Box<RawLegacyFormulaObject>> for LegacyFormula {
-    type Error = DiagnosticReport;
-
-    fn try_from(value: Box<RawLegacyFormulaObject>) -> Result<Self, Self::Error> {
-        (*value).try_into()
-    }
-}
-
-impl TryFrom<RawLegacyFormulaObject> for LegacyFormula {
-    type Error = DiagnosticReport;
-
-    fn try_from(value: RawLegacyFormulaObject) -> Result<Self, Self::Error> {
-        let mut formulas = Vec::new();
-        push_optional_legacy_formula(&mut formulas, value.pattern, Self::Pattern);
-        push_optional_legacy_formula(&mut formulas, value.pattern_regex, Self::PatternRegex);
-        if let Some(patterns) = value.patterns {
-            formulas.push(Self::Patterns(
-                patterns
-                    .into_iter()
-                    .map(TryInto::try_into)
-                    .collect::<Result<Vec<_>, _>>()?,
-            ));
-        }
-        if let Some(pattern_either) = value.pattern_either {
-            formulas.push(Self::PatternEither(
-                pattern_either
-                    .into_iter()
-                    .map(TryInto::try_into)
-                    .collect::<Result<Vec<_>, _>>()?,
-            ));
-        }
-        push_optional_legacy_value_formula(&mut formulas, value.pattern_not, Self::PatternNot)?;
-        push_optional_legacy_value_formula(
-            &mut formulas,
-            value.pattern_inside,
-            Self::PatternInside,
-        )?;
-        push_optional_legacy_value_formula(
-            &mut formulas,
-            value.pattern_not_inside,
-            Self::PatternNotInside,
-        )?;
-        push_optional_legacy_formula(
-            &mut formulas,
-            value.pattern_not_regex,
-            Self::PatternNotRegex,
-        );
-        push_optional_legacy_value_formula(&mut formulas, value.anywhere, Self::Anywhere)?;
-
-        singleton_formula(formulas, |len| match len {
-            0 => schema_error(
-                String::from("legacy formula object is empty"),
-                None,
-                "add a supported legacy operator",
-            ),
-            _ => schema_error(
-                String::from("legacy formula object defines multiple operators"),
-                None,
-                "keep only one operator per legacy object",
-            ),
-        })
-    }
-}
-
-fn push_optional_legacy_formula(
-    formulas: &mut Vec<LegacyFormula>,
-    value: Option<String>,
-    constructor: fn(String) -> LegacyFormula,
-) {
-    if let Some(text) = value {
-        formulas.push(constructor(text));
-    }
-}
-
-fn push_optional_legacy_value_formula(
-    formulas: &mut Vec<LegacyFormula>,
-    value: Option<RawLegacyValue>,
-    constructor: fn(Box<LegacyValue>) -> LegacyFormula,
-) -> Result<(), DiagnosticReport> {
-    if let Some(inner) = value {
-        formulas.push(constructor(Box::new(inner.try_into()?)));
-    }
-    Ok(())
-}
-
-impl TryFrom<RawLegacyValue> for LegacyValue {
-    type Error = DiagnosticReport;
-
-    fn try_from(value: RawLegacyValue) -> Result<Self, Self::Error> {
-        match value {
-            RawLegacyValue::String(text) => Ok(Self::String(text)),
-            RawLegacyValue::Formula(formula) => Ok(Self::Formula(formula.try_into()?)),
-        }
-    }
-}
-
-impl TryFrom<RawMatchFormula> for MatchFormula {
-    type Error = DiagnosticReport;
-
-    fn try_from(value: RawMatchFormula) -> Result<Self, Self::Error> {
-        match value {
-            RawMatchFormula::String(pattern) => Ok(Self::Pattern(pattern)),
-            RawMatchFormula::Object(object) => object.try_into(),
-        }
-    }
-}
-
-impl TryFrom<Box<RawMatchFormulaObject>> for MatchFormula {
-    type Error = DiagnosticReport;
-
-    fn try_from(value: Box<RawMatchFormulaObject>) -> Result<Self, Self::Error> {
-        (*value).try_into()
-    }
-}
-
-/// Converts the operator fields of a `RawMatchFormulaObject` into the core
-/// (undecorated) `MatchFormula` variant.  The caller is responsible for
-/// ensuring exactly one operator field is `Some` before calling this.
-fn build_core_match_formula(
-    value: RawMatchFormulaObject,
-) -> Result<MatchFormula, DiagnosticReport> {
-    if let Some(pattern) = value.pattern {
-        Ok(MatchFormula::PatternObject(pattern))
-    } else if let Some(regex) = value.regex {
-        Ok(MatchFormula::Regex(regex))
-    } else if let Some(all) = value.all {
-        Ok(MatchFormula::All(
-            all.into_iter()
-                .map(TryInto::try_into)
-                .collect::<Result<Vec<_>, _>>()?,
-        ))
-    } else if let Some(any) = value.any {
-        Ok(MatchFormula::Any(
-            any.into_iter()
-                .map(TryInto::try_into)
-                .collect::<Result<Vec<_>, _>>()?,
-        ))
-    } else if let Some(not) = value.not {
-        Ok(MatchFormula::Not(Box::new((*not).try_into()?)))
-    } else if let Some(inside) = value.inside {
-        Ok(MatchFormula::Inside(Box::new((*inside).try_into()?)))
-    } else if let Some(anywhere) = value.anywhere {
-        Ok(MatchFormula::Anywhere(Box::new((*anywhere).try_into()?)))
-    } else {
-        // Safety: caller guarantees operator_count == 1
-        Err(schema_error(
-            String::from("internal error: no operator found despite count check"),
-            None,
-            "please report this bug",
-        ))
-    }
-}
-
-impl TryFrom<RawMatchFormulaObject> for MatchFormula {
-    type Error = DiagnosticReport;
-
-    fn try_from(value: RawMatchFormulaObject) -> Result<Self, Self::Error> {
-        let operator_count = [
-            value.pattern.is_some(),
-            value.regex.is_some(),
-            value.all.is_some(),
-            value.any.is_some(),
-            value.not.is_some(),
-            value.inside.is_some(),
-            value.anywhere.is_some(),
-        ]
-        .iter()
-        .filter(|&&present| present)
-        .count();
-
-        if operator_count == 0 {
-            return Err(schema_error(
-                String::from("match formula object is empty"),
-                None,
-                "add a supported `match` operator",
-            ));
-        }
-
-        if operator_count > 1 {
-            return Err(schema_error(
-                String::from("match formula object defines multiple operators"),
-                None,
-                "keep only one operator per match object",
-            ));
-        }
-
-        let where_ = value.where_clauses.clone().unwrap_or_default();
-        let as_name = value.as_name.clone();
-        let fix = value.fix.clone();
-        let core = build_core_match_formula(value)?;
-        Ok(Self::decorated(core, where_, as_name, fix))
-    }
 }
